SQLEXEC-54: Add Rule Access Control for Tenant, Ruleset, and Schema APIs

+++ backend/app/api/routers/schema_manager.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 
 from api.core.services.schema.schema_manager_service import SchemaManagerService
 from model.requests.schema_manager.add_schema_request import AddSchemaRequest
@@ -10,17 +10,37 @@ from utils.jwt_utils import authenticate_admin_session
 router = APIRouter()
 
 @router.post("/{tenant_id}/schemas/")
-async def add_schema(tenant_id: str, schema_request: AddSchemaRequest, _: AdminSessionData = Depends(authenticate_admin_session)):
+async def add_schema(tenant_id: str, schema_request: AddSchemaRequest, admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Schema Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await SchemaManagerService.add_schema(tenant_id=tenant_id, schema_request=schema_request)
 
 @router.get("/{tenant_id}/schemas/{schema_name}")
-async def get_schema(tenant_id: str, schema_name: str,  _: AdminSessionData = Depends(authenticate_admin_session)):
+async def get_schema(tenant_id: str, schema_name: str,  admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Schema Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await SchemaManagerService.get_schema(tenant_id=tenant_id, schema_name=schema_name)
 
 @router.put("/{tenant_id}/schemas/{schema_name}")
-async def update_schema(tenant_id: str, schema_name: str, schema_request: UpdateSchemaRequest,  _: AdminSessionData = Depends(authenticate_admin_session)):
+async def update_schema(tenant_id: str, schema_name: str, schema_request: UpdateSchemaRequest,  admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Schema Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await SchemaManagerService.update_schema(tenant_id=tenant_id, schema_name=schema_name, update_schema_request=schema_request)
 
 @router.delete("/{tenant_id}/schemas/{schema_name}")
-async def delete_schema(tenant_id: str, schema_name: str,  _: AdminSessionData = Depends(authenticate_admin_session)):
+async def delete_schema(tenant_id: str, schema_name: str,  admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Schema Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await SchemaManagerService.delete_schema(tenant_id=tenant_id, schema_name=schema_name)
